@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.3.4+16
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
   flutter:
@@ -32,9 +32,9 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-  intl_phone_number_input: ^0.7.0+2
+  intl_phone_number_input: ^0.7.3
   url_launcher: ^6.0.17
-  package_info_plus: ^1.4.0
+  package_info_plus: ^3.1.2
   permission_handler: ^10.2.0
   contacts_service: ^0.6.3
   get_it: ^6.0.0
